CodeAgentTask/requirements.txt
openai>=1.35.0
python-dotenv>=1.0.1
pytest>=8.0.0
streamlit